% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_ibisma-package.R
\docType{package}
\name{painel_ibisma-package}
\alias{painel_ibisma}
\alias{painel_ibisma-package}
\title{painel_ibisma: Painel do Indice Brasileiro de Inseguranca em Saude Materna}
\description{
Pacote que implementa o painel interativo do IBISMA com Shiny e golem.
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

Authors:
\itemize{
  \item firstname lastname \email{your@email.com}
}

}
\keyword{internal}
