% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/painel_ibisma_v4-package.R
\docType{package}
\name{painel_ibisma_v4-package}
\alias{painel_ibisma_v4}
\alias{painel_ibisma_v4-package}
\title{painel_ibisma_v4: Painel do Indice Brasileiro de Inseguranca em Saude Materna}
\description{
Pacote que implementa o painel interativo do IBISMA com Shiny e golem.
}
\author{
\strong{Maintainer}: firstname lastname \email{your@email.com}

}
\keyword{internal}
